Moved ExtUITabStrip to UI.Helpers folder and namespace.
