+++ TLM/TLM/State/OptionsTabs/OptionsGameplayTab.cs
@@ -4,6 +4,7 @@ namespace TrafficManager.State {
     using ICities;
     using Manager.Impl;
     using UI;
+    using UI.Helpers;
     using UnityEngine;
 
     public static class OptionsGameplayTab {
